Revert "Store the paged KV cache as FP16"
This reverts commit e2aac14cdfac3abd096dcafa3ecf8e593061245e.

+++ crates/hi-cuda/src/lib.rs
@@ -533,8 +533,7 @@ impl CudaPagedKvCacheManager {
             .checked_mul(kv_heads)
             .and_then(|value| value.checked_mul(qk_head_dim.checked_add(v_head_dim)?))
             .and_then(|value| value.checked_mul(page_size))
-            // KV pages are stored as f16 (kv_t = __half in kernels.cu).
-            .and_then(|value| value.checked_mul(std::mem::size_of::<u16>()))
+            .and_then(|value| value.checked_mul(std::mem::size_of::<f32>()))
             .context("CUDA paged KV page byte count overflows usize")?;
         let bytes_total = bytes_per_page
             .checked_mul(pages_total)
